feature: Add Day 14 implementation
- Added Day 14 implementation: quickly found scaling to be a problem :rofl:

To solve this, instead look at each pair value as mappings. This map expands to however many steps there are. I got stuck on the end bit when I had to find the largest difference of said map. Because of the map being pairs of the actual polymer string, you end up with doubled the count for each character. This took me quite some time to spot, unfortunately... 😓