feature: Add Day 15 implementation
- Added Day 15 implementation: I at first looked to use OOP to solve this problem but in the end found an algorithmic approach to solve it.

To solve this, I looked at building the path by summing the risks in a grid. At the end, the position of [0, 0] will have the sum of the minimum risk to go from bottom-right(end) to top-left(start). The drawback to this is the slowness as graph algorithms aren't used. Each time the grid of sum of risks is modified, the entire grid must be traced back again to check for potential minimisation.

Part II on the other hand was quite straightforward. I explained this in the implementation.