feature: Add Day 16 implementation
- Added Day 16 implementation: I first looked to solve the problem without OOP. After being defeated horribly by the sheer wall of task description, I gave up and resulted to OOP.

To solve part I, I had to convert the hex string to binary. I parsed the hex to denary then back to binary, left padding with 0s to a length of 4. Next, I had to separate all packets within a packet. If you think about this as a tree, it ends up being a binary tree with literal packets(type ID 4) as the child node of an operator packet(type ID != 4). This is tied into part II where we have to apply various functions depending on the type ID.

Although part II was simple to implement, as I didn't read the task description well enough, I completely messed up handling the literal packets. I read the whole literal packet binary string and converted it to denary, which is false. Every sub-string of the literal has 5 bits, with the first one denoting whether it is the end of the literal or not. After digging through all my methods I found this fix and finally finished the solution... This was probably the most challenging one I have found so far, hence why I took more than a day, but hey-o! 💯