feature: Add Day 17 implementation
- Added Day 17 implementation: Part II was more of a headache than it should've been... 😭

Solving part I required the use of the arithmetic sum equation, explained further in the code.

Solving part II was simply checking bounds and each step of the velocities. However, I messed up reading in the input and also messed up somewhere regarding the calculation and velocity changes. This was a huge headache and nearly made me miss my commit!