Rewrite project structure
Figured it'd be better to rewrite this now than muddle through when I get to the point. There's some hard-coding required that can be easily resolved via using real-term dates, but I don't like that as I'm not guaranteed to finish a solution per-day/even in the year.

Can eventually use a config file instead.