Add final implementation of Day 3
Day 3 was simple, but I did spend a lot of time and effort working out how what to map.

The biggest performance hit was the use of nested loops to calculate intersecting characters. There is support for set operation(containsAll()?) but I just opted for the loop. Funnily regex replacement takes longer than a brute-force loop.

Task 2 was straightforward too - you can reuse part. But I like to separate both implementations so I don't use part 1 for 2. Part 1 calculates the score(if you did it that way) for the 1st and 2nd rucksack, so you merely need to calculate the additional 3rd rucksack and sum it.