Add final implementation of Day4
Day 4 wasn't difficult but the bounds made me think twice; I also wasn't able to finish it on the day as something else came up.

Idea is that in task 1, we just need to check if either contained the other e.g. a[0] <= b[0] && a[1] >= b[1] and vice versa.

Task 2 is similar, but instead we need to check that a value can encompass the other e.g. 6-10, 5-20; in which we have to check that a[0] >= b[0] && a[0] <= b[1] or a[1] >= b[0] && a[1] <= b[1] - case of 15-20, 9-21.