Add final implementation of Day5
Day 5 at first seemed complicated due to the input parsing - but is really just a stack-based problem. You could realistically do this with an array too; but it would require more bounds checking. I was looking to do an array but as I started late I decided to go with a stack-based approach instead.

With task 2, you could instead use a queue, two stacks, or again - array bounds. I opted for the two stacks because of time; pushing to a tmp stack and using that to push the elements again just reverses it.